examples/boids: [Enh] Move implementations to .cpp fles
So it is not header only.
